Remove app.yml v1
