+++ app/lib/metadata.py
@@ -7,7 +7,7 @@ import yaml
 
 from lib.composegenerator.next.stage1 import createCleanConfigFromV3
 from lib.composegenerator.v2.networking import getMainContainer
-from lib.composegenerator.v1.networking import getFreePort
+from lib.composegenerator.shared.networking import getFreePort
 from lib.entropy import deriveEntropy
 from typing import List
 import json
